+++ artemis-gwt-test/src/test/java/com/artemis/link/EntityLinkManagerTest.java
@@ -11,6 +11,50 @@ public class EntityLinkManagerTest extends GWTTestCase {
 		return "com.ArtemisTest";
 	}
 
+	public void test_unilink_kill_empty_link() {
+		World world = new World(new WorldConfiguration()
+			.setSystem(EntityLinkManager.class));
+
+		final int otherA = world.create();
+		final int e = world.create();
+
+		ComponentMapper<EntityLink> mapper = world.getMapper(EntityLink.class);
+
+		EntityLinkManager elm = world.getSystem(EntityLinkManager.class);
+		elm.register(EntityLink.class, "otherId", new EmptyLinkListener(e, otherA, -1));
+
+		// establish empty link
+		mapper.create(e);
+
+		world.process();
+
+		// kill link
+		world.delete(e);
+		world.process();
+	}
+
+	public void test_multilink_kill_empty_link() {
+		World world = new World(new WorldConfiguration()
+			.setSystem(EntityLinkManager.class));
+
+		final int otherA = world.create();
+		final int e = world.create();
+
+		ComponentMapper<MultiLinkCheckAll> mapper = world.getMapper(MultiLinkCheckAll.class);
+
+		EntityLinkManager elm = world.getSystem(EntityLinkManager.class);
+		elm.register(EntityLink.class, "otherId", new EmptyLinkListener(e, otherA, -1));
+
+		// establish empty link
+		mapper.create(e);
+
+		world.process();
+
+		// kill link
+		world.delete(e);
+		world.process();
+	}
+
 	public void test_unilink_explicit_field_int_test() {
 		World world = new World(new WorldConfiguration()
 			.setSystem(EntityLinkManager.class));
@@ -433,6 +477,44 @@ public class EntityLinkManagerTest extends GWTTestCase {
 		@Override
 		public void onTargetChanged(int sourceId, int targetId, int oldTargetId) {
 			fail();
+		}
+	}
+
+
+	public static class EmptyLinkListener implements LinkListener {
+		private final int e;
+		private final int otherA;
+		private final int otherB;
+
+		public EmptyLinkListener(int e, int otherA, int otherB) {
+			this.e = e;
+			this.otherA = otherA;
+			this.otherB = otherB;
+		}
+
+		@Override
+		public void onLinkEstablished(int sourceId, int targetId) {
+			assertEquals(sourceId, e);
+			assertEquals(targetId, otherA);
+		}
+
+		@Override
+		public void onLinkKilled(int sourceId, int targetId) {
+			assertEquals(sourceId, e);
+			assertEquals(targetId, -1);
+		}
+
+		@Override
+		public void onTargetDead(int sourceId, int deadTargetId) {
+			assertEquals(sourceId, e);
+			assertEquals(deadTargetId, otherB);
+		}
+
+		@Override
+		public void onTargetChanged(int sourceId, int targetId, int oldTargetId) {
+			assertEquals(sourceId, e);
+			assertEquals(targetId, otherB);
+			assertEquals(oldTargetId, otherA);
 		}
 	}
 }