+++ artemis-test/src/test/java/com/artemis/link/EntityLinkManagerTest.java
@@ -1,12 +1,64 @@
 package com.artemis.link;
 
 import com.artemis.*;
-import com.artemis.component.link.*;
+import com.artemis.annotations.EntityId;
+import com.artemis.annotations.LinkPolicy;
+import com.artemis.utils.Bag;
+import com.artemis.utils.IntBag;
 import org.junit.Test;
 
+import static com.artemis.annotations.LinkPolicy.Policy.CHECK_SOURCE;
+import static com.artemis.annotations.LinkPolicy.Policy.CHECK_SOURCE_AND_TARGETS;
+import static com.artemis.annotations.LinkPolicy.Policy.SKIP;
 import static org.junit.Assert.*;
 
 public class EntityLinkManagerTest {
+
+	@Test
+	public void unilink_kill_empty_link_test() {
+		World world = new World(new WorldConfiguration()
+			.setSystem(EntityLinkManager.class));
+
+		final int otherA = world.create();
+		final int e = world.create();
+
+		ComponentMapper<EntityLink> mapper = world.getMapper(EntityLink.class);
+
+		EntityLinkManager elm = world.getSystem(EntityLinkManager.class);
+		elm.register(EntityLink.class, "otherId", new EmptyLinkListener(e, otherA, -1));
+
+		// establish empty link
+		mapper.create(e);
+
+		world.process();
+
+		// kill link
+		world.delete(e);
+		world.process();
+	}
+
+	@Test
+	public void multilink_kill_empty_link_test() {
+		World world = new World(new WorldConfiguration()
+			.setSystem(EntityLinkManager.class));
+
+		final int otherA = world.create();
+		final int e = world.create();
+
+		ComponentMapper<MultiLinkCheckAll> mapper = world.getMapper(MultiLinkCheckAll.class);
+
+		EntityLinkManager elm = world.getSystem(EntityLinkManager.class);
+		elm.register(EntityLink.class, "otherId", new EmptyLinkListener(e, otherA, -1));
+
+		// establish empty link
+		mapper.create(e);
+
+		world.process();
+
+		// kill link
+		world.delete(e);
+		world.process();
+	}
 
 	@Test
 	public void unilink_explicit_field_int_test() {
@@ -409,6 +461,50 @@ public class EntityLinkManagerTest {
 		world.process();
 	}
 
+	public static class EntityLink extends Component {
+		@EntityId
+		public int otherId = -1;
+		public int nothingHere;
+	}
+
+	public static class EntityLinkB extends Component {
+		public Entity other;
+		public int nothingHere;
+	}
+
+	public static class EntityLinkC extends Component {
+		@LinkPolicy(CHECK_SOURCE)
+		public Entity other;
+	}
+
+	public static class EntityLinkSkip extends Component {
+		@LinkPolicy(SKIP)
+		public Entity other;
+	}
+
+	public static class MultiLinkSkip extends Component {
+		@LinkPolicy(SKIP)
+		public Bag<Entity> other = new Bag<Entity>();
+	}
+
+	public static class MultiLinkSkipTargetCheck extends Component {
+		public Bag<Entity> other = new Bag<Entity>();
+	}
+
+	public static class MultiLinkCheckAll extends Component {
+		@EntityId @LinkPolicy(CHECK_SOURCE_AND_TARGETS)
+		public IntBag other = new IntBag();
+	}
+
+
+	public static class MuchOfEverything extends Component {
+		@EntityId public IntBag intIds = new IntBag();
+		@EntityId public int otherId;
+		public Entity e;
+		public Bag<Entity> entities = new Bag<Entity>();
+		public int notMe;
+	}
+
 	private static class MyLinkListener implements LinkListener {
 		private final int e;
 		private final int otherA;
@@ -430,6 +526,43 @@ public class EntityLinkManagerTest {
 		public void onLinkKilled(int sourceId, int targetId) {
 			assertEquals(sourceId, e);
 			assertEquals(targetId, otherA);
+		}
+
+		@Override
+		public void onTargetDead(int sourceId, int deadTargetId) {
+			assertEquals(sourceId, e);
+			assertEquals(deadTargetId, otherB);
+		}
+
+		@Override
+		public void onTargetChanged(int sourceId, int targetId, int oldTargetId) {
+			assertEquals(sourceId, e);
+			assertEquals(targetId, otherB);
+			assertEquals(oldTargetId, otherA);
+		}
+	}
+
+	private static class EmptyLinkListener implements LinkListener {
+		private final int e;
+		private final int otherA;
+		private final int otherB;
+
+		public EmptyLinkListener(int e, int otherA, int otherB) {
+			this.e = e;
+			this.otherA = otherA;
+			this.otherB = otherB;
+		}
+
+		@Override
+		public void onLinkEstablished(int sourceId, int targetId) {
+			assertEquals(sourceId, e);
+			assertEquals(targetId, otherA);
+		}
+
+		@Override
+		public void onLinkKilled(int sourceId, int targetId) {
+			assertEquals(sourceId, e);
+			assertEquals(targetId, -1);
 		}
 
 		@Override