+++ artemis/src/main/java/com/artemis/link/UniLinkSite.java
@@ -23,24 +23,29 @@ class UniLinkSite extends LinkSite {
 		// -1 == not linked
 		int target = fieldMutator.read(mapper.get(id), field);
 		if (target != -1 && !activeEntityIds.unsafeGet(target)) {
+			// target is dead or linked field is set to null/-1
 			target = -1;
 			fieldMutator.write(target, mapper.get(id), field);
 		}
 
 		int oldTarget = sourceToTarget.get(id);
 		if (target != oldTarget) {
-			sourceToTarget.set(id, target);
-			if (oldTarget == -1) {
-				if (listener != null) listener.onLinkEstablished(id, target);
-			} else {
-				if (listener != null) {
-					if (target != -1) {
-						listener.onTargetChanged(id, target, oldTarget);
-					} else {
-						listener.onTargetDead(id, oldTarget);
-					}
-				}
+			if (listener != null) {
+				fireLinkListener(id, target);
 			}
+
+			sourceToTarget.set(id, target);
+		}
+	}
+
+	private void fireLinkListener(int id, int target) {
+		int oldTarget = sourceToTarget.get(id);
+		if (oldTarget == -1) {
+			listener.onLinkEstablished(id, target);
+		} else if (target != -1) {
+			listener.onTargetChanged(id, target, oldTarget);
+		} else {
+			listener.onTargetDead(id, oldTarget);
 		}
 	}
 
@@ -58,9 +63,10 @@ class UniLinkSite extends LinkSite {
 			? sourceToTarget.get(id)
 			: -1;
 
-		if (target != -1) {
+		if (target != -1)
 			sourceToTarget.set(id, -1);
-			if (listener != null) listener.onLinkKilled(id, target);
-		}
+
+		if (listener != null)
+			listener.onLinkKilled(id, target);
 	}
 }