+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Version: 2.0.0-SNAPSHOT
 - **Fix**: `IntBag::get(index)`, ArrayIndexOutOfBoundsException reported size of backing array, not logical size.
 - **Fix**: Possible IOOB exception when registering more than 64 systems.
+- **Fix**: EntityLinkManager/LinkListener::onLinkKilled never fired for unestablished links.
 
 
 #### Version: 2.0.0-RC5 - 2016-07-23